lib/server/requirements.txt
Flask
flask-cors
feedparser
gunicorn
google-cloud-firestore
python-dotenv
openai
fpdf
reportlab


